Move dial funcs to dialer pkg
This reduces shim size from 30mb to 18mb
