+++ dialer/dialer.go
@@ -1,6 +1,7 @@
-package containerd
+package dialer
 
 import (
+	"fmt"
 	"net"
 	"time"
 
@@ -10,6 +11,12 @@ import (
 type dialResult struct {
 	c   net.Conn
 	err error
+}
+
+// DialAddress returns the address with unix:// prepended to the
+// provided address
+func DialAddress(address string) string {
+	return fmt.Sprintf("unix://%s", address)
 }
 
 // Dialer returns a GRPC net.Conn connected to the provided address
