+++ dialer/dialer_unix.go
@@ -1,9 +1,8 @@
 // +build !windows
 
-package containerd
+package dialer
 
 import (
-	"fmt"
 	"net"
 	"os"
 	"strings"
@@ -27,10 +26,4 @@ func isNoent(err error) bool {
 func dialer(address string, timeout time.Duration) (net.Conn, error) {
 	address = strings.TrimPrefix(address, "unix://")
 	return net.DialTimeout("unix", address, timeout)
-}
-
-// DialAddress returns the address with unix:// prepended to the
-// provided address
-func DialAddress(address string) string {
-	return fmt.Sprintf("unix://%s", address)
 }
